Turn registered contracts into a call table on the server, create a function for deserializing message headers

+++ src/message-headers/request-header.zig
@@ -4,7 +4,7 @@ pub const RequestHeader = struct {
     version: u8,
     msg_type: MessageType,
     request_id: u32,
-    contract_id: u32,
+    contract_id: u16,
     method_id: u16,
     payload_len: u32,
 };
